# [Perf] SM100 FP8 GEMM Optimizations after cutlass_profiler