Additional performance optimizations after #19566

Tune CUTLASS configs for M <= 256 using cutlass_profiler insights.
The PR allows to get from 1.12x to 1.5x speedup for certain NxK pairs without affecting performance for the other NxK.

## Cutlass profiler results. 
We pick the best configuration that has best performance across different NxK pairs for static M.
cta_m,   cta_n,   cta_k,   cluster_m,   cluster_n,   cluster_k are the parameters. Rank represents the rank of the performance parameters set for a single NxK in cutlass_profiler comparison, GFPLOPs - performance results in the corresponding benchmark for all NxK pairs.
M=16

|   cta_m |   cta_n |   cta_k |   cluster_m |   cluster_n |   cluster_k | Rank                                                                                                                            | GFLOPs                                                          |
|--------:|--------:|--------:|------------:|------------:|------------:|:--------------------------------------------------------------------------------------------------------------------------------------------------------|:-----------------------------------------------------------------------------|
|      64 |      64 |     128 |           1 |           4 |           1 | ['16x2560x8192:3', '16x28672x4096:1', '16x4096x14336:5', '16x14336x8192:8', '16x4096x4096:8', '16x8192x7168:2', '16x6144x4096:2', '16x8192x2048:1']     | [54489.5, 146659.0, 101793.0, 139143.0, 52288.0, 114496.0, 77335.6, 54581.4] |
|     128 |     128 |     128 |           2 |           1 |           1 | ['16x2560x8192:6', '16x28672x4096:11', '16x4096x14336:4', '16x14336x8192:6', '16x4096x4096:7', '16x8192x7168:5', '16x6144x4096:8', '16x8192x2048:4']    | [54301.8, 133538.0, 101793.0, 140305.0, 52292.9, 111033.0, 66998.4, 52344.8] |
|      64 |      64 |     128 |           1 |           2 |           1 | ['16x2560x8192:2', '16x28672x4096:9', '16x4096x14336:3', '16x14336x8192:15', '16x4096x4096:1', '16x8192x7168:3', '16x6144x4096:5', '16x8192x2048:8']    | [54492.3, 135616.0, 101841.0, 136116.0, 52382.7, 113502.0, 74268.9, 52338.3] |
|      64 |      64 |     256 |           1 |           4 |           1 | ['16x2560x8192:14', '16x28672x4096:4', '16x4096x14336:13', '16x14336x8192:10', '16x4096x4096:5', '16x8192x7168:1', '16x6144x4096:3', '16x8192x2048:2']  | [50150.8, 142839.0, 99246.8, 137627.0, 52317.3, 114500.0, 76734.4, 54349.9]  |
|      64 |      64 |     128 |           1 |           8 |           1 | ['16x2560x8192:5', '16x28672x4096:3', '16x4096x14336:6', '16x14336x8192:9', '16x4096x4096:2', '16x8192x7168:13', '16x6144x4096:1', '16x8192x2048:14']   | [54485.2, 145789.0, 101781.0, 137939.0, 52336.9, 96850.9, 77881.2, 52295.9]  |
|      64 |      64 |     128 |           1 |           1 |           1 | ['16x2560x8192:1', '16x28672x4096:15', '16x4096x14336:1', '16x14336x8192:4', '16x4096x4096:10', '16x8192x7168:11', '16x6144x4096:10', '16x8192x2048:6'] | [54492.3, 125416.0, 104745.0, 141677.0, 52281.5, 101680.0, 65444.6, 52341.5] |
|     128 |     128 |     128 |           2 |           2 |           1 | ['16x2560x8192:9', '16x28672x4096:13', '16x4096x14336:9', '16x14336x8192:11', '16x4096x4096:4', '16x8192x7168:10', '16x6144x4096:9', '16x8192x2048:10'] | [52898.1, 130875.0, 101497.0, 137532.0, 52330.4, 101851.0, 65446.3, 52310.5] |

M=32

|   cta_m |   cta_n |   cta_k |   cluster_m |   cluster_n |   cluster_k | Rank                                                                                                                            | GFLOPs                                                              |
|--------:|--------:|--------:|------------:|------------:|------------:|:--------------------------------------------------------------------------------------------------------------------------------------------------------|:---------------------------------------------------------------------------------|
|      64 |      64 |     128 |           1 |           1 |           1 | ['32x2560x8192:1', '32x4096x4096:1', '32x14336x8192:4', '32x8192x7168:1', '32x4096x14336:1', '32x8192x2048:2', '32x28672x4096:2', '32x6144x4096:1']     | [109044.0, 104880.0, 281366.0, 255430.0, 205877.0, 106808.0, 297854.0, 157006.0] |
|      64 |      64 |     128 |           1 |           2 |           1 | ['32x2560x8192:3', '32x4096x4096:2', '32x14336x8192:3', '32x8192x7168:4', '32x4096x14336:5', '32x8192x2048:8', '32x28672x4096:6', '32x6144x4096:3']     | [108976.0, 104680.0, 281896.0, 232749.0, 203572.0, 104680.0, 287400.0, 156679.0] |
|      64 |      64 |     128 |           1 |           4 |           1 | ['32x2560x8192:4', '32x4096x4096:9', '32x14336x8192:9', '32x8192x7168:2', '32x4096x14336:3', '32x8192x2048:1', '32x28672x4096:3', '32x6144x4096:2']     | [108973.0, 104583.0, 277114.0, 235758.0, 203604.0, 109365.0, 297120.0, 156849.0] |
|      64 |      64 |     256 |           1 |           1 |           1 | ['32x2560x8192:7', '32x4096x4096:8', '32x14336x8192:8', '32x8192x7168:3', '32x4096x14336:7', '32x8192x2048:4', '32x28672x4096:13', '32x6144x4096:5']    | [107446.0, 104651.0, 277121.0, 234253.0, 202564.0, 104794.0, 282269.0, 155603.0] |
|     128 |     128 |     128 |           2 |           1 |           1 | ['32x2560x8192:5', '32x4096x4096:6', '32x14336x8192:7', '32x8192x7168:7', '32x4096x14336:6', '32x8192x2048:5', '32x28672x4096:12', '32x6144x4096:11']   | [108418.0, 104661.0, 279744.0, 229269.0, 203565.0, 104784.0, 282374.0, 151110.0] |
|      64 |      64 |     256 |           1 |           4 |           1 | ['32x2560x8192:14', '32x4096x4096:10', '32x14336x8192:11', '32x8192x7168:5', '32x4096x14336:13', '32x8192x2048:3', '32x28672x4096:9', '32x6144x4096:6'] | [98946.0, 104579.0, 274213.0, 229839.0, 194072.0, 105838.0, 283253.0, 154406.0]  |
|     128 |     128 |     128 |           2 |           2 |           1 | ['32x2560x8192:9', '32x4096x4096:13', '32x14336x8192:14', '32x8192x7168:8', '32x4096x14336:8', '32x8192x2048:11', '32x28672x4096:8', '32x6144x4096:10'] | [104131.0, 104573.0, 272192.0, 228992.0, 201892.0, 104589.0, 284123.0, 151570.0] |

M=64
|   cta_m |   cta_n |   cta_k |   cluster_m |   cluster_n |   cluster_k | Rank                                                                                                                           | GFLOPs                                                              |
|--------:|--------:|--------:|------------:|------------:|------------:|:-------------------------------------------------------------------------------------------------------------------------------------------------------|:---------------------------------------------------------------------------------|
|      64 |      64 |     128 |           1 |           1 |           1 | ['64x28672x4096:5', '64x14336x8192:2', '64x2560x8192:1', '64x4096x4096:1', '64x6144x4096:1', '64x8192x7168:1', '64x4096x14336:1', '64x8192x2048:1']    | [570911.0, 563819.0, 218094.0, 209544.0, 313699.0, 495003.0, 407603.0, 260269.0] |
|      64 |      64 |     128 |           1 |           4 |           1 | ['64x28672x4096:2', '64x14336x8192:3', '64x2560x8192:4', '64x4096x4096:7', '64x6144x4096:2', '64x8192x7168:2', '64x4096x14336:4', '64x8192x2048:2']    | [581315.0, 561891.0, 217901.0, 209328.0, 313679.0, 468488.0, 407144.0, 241908.0] |
|      64 |      64 |     128 |           1 |           2 |           1 | ['64x28672x4096:8', '64x14336x8192:7', '64x2560x8192:2', '64x4096x4096:2', '64x6144x4096:4', '64x8192x7168:3', '64x4096x14336:3', '64x8192x2048:4']    | [564532.0, 557186.0, 217975.0, 209518.0, 313084.0, 466719.0, 407186.0, 228778.0] |
|      64 |      64 |     256 |           1 |           1 |           1 | ['64x28672x4096:10', '64x14336x8192:8', '64x2560x8192:6', '64x4096x4096:4', '64x6144x4096:5', '64x8192x7168:4', '64x4096x14336:7', '64x8192x2048:5']   | [563699.0, 548291.0, 213515.0, 209335.0, 311197.0, 465213.0, 403999.0, 227815.0] |
|      64 |      64 |     256 |           1 |           2 |           1 | ['64x28672x4096:11', '64x14336x8192:13', '64x2560x8192:8', '64x4096x4096:5', '64x6144x4096:6', '64x8192x7168:5', '64x4096x14336:9', '64x8192x2048:3']  | [563246.0, 543370.0, 205306.0, 209328.0, 308140.0, 459435.0, 394017.0, 229262.0] |
|     128 |     128 |     128 |           2 |           1 |           1 | ['64x28672x4096:14', '64x14336x8192:9', '64x2560x8192:7', '64x4096x4096:6', '64x6144x4096:9', '64x8192x7168:7', '64x4096x14336:5', '64x8192x2048:7']   | [562585.0, 547882.0, 209893.0, 209328.0, 301623.0, 458252.0, 405590.0, 217406.0] |
|     128 |     128 |     128 |           2 |           2 |           1 | ['64x28672x4096:7', '64x14336x8192:10', '64x2560x8192:10', '64x4096x4096:9', '64x6144x4096:10', '64x8192x7168:8', '64x4096x14336:8', '64x8192x2048:8'] | [564539.0, 547824.0, 201562.0, 209322.0, 297717.0, 457984.0, 398070.0, 212527.0] |

M=128
|   cta_m |   cta_n |   cta_k |   cluster_m |   cluster_n |   cluster_k | Rank                                                                                                                                   | GFLOPs                                                                |
|--------:|--------:|--------:|------------:|------------:|------------:|:---------------------------------------------------------------------------------------------------------------------------------------------------------------|:-----------------------------------------------------------------------------------|
|     128 |     128 |     128 |           2 |           1 |           1 | ['128x2560x8192:8', '128x4096x4096:9', '128x14336x8192:4', '128x8192x7168:1', '128x8192x2048:1', '128x28672x4096:3', '128x6144x4096:2', '128x4096x14336:5']    | [415092.0, 418265.0, 1065420.0, 916540.0, 426784.0, 1097860.0, 577819.0, 801383.0] |
|     128 |     128 |     128 |           2 |           2 |           1 | ['128x2560x8192:10', '128x4096x4096:8', '128x14336x8192:2', '128x8192x7168:2', '128x8192x2048:2', '128x28672x4096:2', '128x6144x4096:1', '128x4096x14336:6']   | [390743.0, 418278.0, 1081600.0, 915932.0, 420450.0, 1126510.0, 582752.0, 770648.0] |
|      64 |      64 |     128 |           1 |           1 |           1 | ['128x2560x8192:1', '128x4096x4096:1', '128x14336x8192:13', '128x8192x7168:5', '128x8192x2048:5', '128x28672x4096:8', '128x6144x4096:5', '128x4096x14336:1']   | [435950.0, 418656.0, 907425.0, 814372.0, 418785.0, 1001930.0, 523095.0, 814753.0]  |
|      64 |      64 |     128 |           2 |           1 |           1 | ['128x2560x8192:3', '128x4096x4096:7', '128x14336x8192:15', '128x8192x7168:9', '128x8192x2048:3', '128x28672x4096:9', '128x6144x4096:4', '128x4096x14336:2']   | [433685.0, 418291.0, 891168.0, 814302.0, 418837.0, 978594.0, 523163.0, 814400.0]   |
|      64 |      64 |     128 |           1 |           2 |           1 | ['128x2560x8192:2', '128x4096x4096:10', '128x14336x8192:12', '128x8192x7168:8', '128x8192x2048:4', '128x28672x4096:10', '128x6144x4096:8', '128x4096x14336:3'] | [435825.0, 418265.0, 909419.0, 814330.0, 418811.0, 977952.0, 520539.0, 814287.0]   |

Kernel benchmarks using #17126 on B200.
`python benchmarks/kernels/bench_fp8_gemm.py --model meta-llama/Llama-3.1-8B-Instruct --tp-sizes 1`
and `python benchmarks/kernels/bench_fp8_gemm.py --model meta-llama/Llama-3.3-70B-Instruct --tp-sizes 4`

## meta-llama/Llama-3.1-8B-Instruct
### N=4096 K=14336

| batch_size | fp8-tensor-w-tensor-a-noquant (Before) | fp8-tensor-w-tensor-a-noquant (After) | speedup |
|-----------:|---------------------------------------:|--------------------------------------:|----------:|
|          1 |                                  10.22 |                                 10.08 |      0.99 |
|         16 |                                 162.61 |                                161.47 |      0.99 |
|         32 |                                 323.89 |                                321.19 |      0.99 |
|         64 |                                 642.23 |                                637.38 |      0.99 |
|        128 |                                1225.21 |                               1234.45 |      1.01 |
|        192 |                                1597.06 |                               1649.77 |      1.03 |
|        256 |                                2059.40 |                               2242.74 |      1.09 |

---

## meta-llama/Llama-3.1-8B-Instruct
### N=4096 K=4096

| batch_size | fp8-tensor-w-tensor-a-noquant (Before) | fp8-tensor-w-tensor-a-noquant (After) | speedup |
|-----------:|---------------------------------------:|--------------------------------------:|----------:|
|          1 |                                   5.86 |                                  5.69 |      0.97 |
|         16 |                                  93.22 |                                 91.95 |      0.99 |
|         32 |                                 185.81 |                                188.96 |      1.02 |
|         64 |                                 367.71 |                                373.51 |      1.02 |
|        128 |                                 703.56 |                                708.25 |      1.01 |
|        192 |                                 925.66 |                                989.09 |      1.07 |
|        256 |                                1246.88 |                               1335.40 |      1.07 |

---

## meta-llama/Llama-3.1-8B-Instruct
### N=6144 K=4096

| batch_size | fp8-tensor-w-tensor-a-noquant (Before) | fp8-tensor-w-tensor-a-noquant (After) | speedup |
|-----------:|---------------------------------------:|--------------------------------------:|----------:|
|          1 |                                   8.41 |                                  7.86 |      0.93 |
|         16 |                                 137.74 |                                128.51 |      0.93 |
|         32 |                                 271.65 |                                278.66 |      1.03 |
|         64 |                                 533.76 |                                553.45 |      1.04 |
|        128 |                                1019.67 |                               1041.53 |      1.02 |
|        192 |                                1019.66 |                               1024.30 |      1.00 |
|        256 |                                1373.78 |                               1354.85 |      0.99 |

---

## meta-llama/Llama-3.1-8B-Instruct
### N=28672 K=4096

| batch_size | fp8-tensor-w-tensor-a-noquant (Before) | fp8-tensor-w-tensor-a-noquant (After) | speedup |
|-----------:|---------------------------------------:|--------------------------------------:|----------:|
|          1 |                                  10.10 |                                 10.62 |      1.05 |
|         16 |                                 159.61 |                                168.90 |      1.06 |
|         32 |                                 314.99 |                                332.05 |      1.05 |
|         64 |                                 619.51 |                                648.96 |      1.05 |
|        128 |                                1182.28 |                               1192.15 |      1.01 |
|        192 |                                1332.86 |                               1320.79 |      0.99 |
|        256 |                                1663.22 |                               1630.00 |      0.98 |

---

## meta-llama/Llama-3.3-70B-Instruct
### N=14336 K=8192

| batch_size | fp8-tensor-w-tensor-a-noquant (Before) | fp8-tensor-w-tensor-a-noquant (After) | speedup |
|-----------:|---------------------------------------:|--------------------------------------:|----------:|
|          1 |                                   9.91 |                                  9.41 |      0.95 |
|         16 |                                 157.88 |                                166.88 |      1.06 |
|         32 |                                 313.35 |                                333.66 |      1.06 |
|         64 |                                 618.64 |                                622.06 |      1.01 |
|        128 |                                1170.04 |                               1195.33 |      1.02 |
|        192 |                                1280.06 |                               1238.12 |      0.97 |
|        256 |                                1610.38 |                               1620.58 |      1.01 |

---

## meta-llama/Llama-3.3-70B-Instruct
### N=2560 K=8192

| batch_size | fp8-tensor-w-tensor-a-noquant (Before) | fp8-tensor-w-tensor-a-noquant (After) | speedup |
|-----------:|---------------------------------------:|--------------------------------------:|----------:|
|          1 |                                   5.27 |                                  5.21 |      0.99 |
|         16 |                                  83.95 |                                 83.13 |      0.99 |
|         32 |                                 167.41 |                                165.62 |      0.99 |
|         64 |                                 332.07 |                                328.48 |      0.99 |
|        128 |                                 635.35 |                                634.71 |      1.00 |
|        192 |                                 938.49 |                                935.84 |      1.00 |
|        256 |                                1247.44 |                               1245.94 |      1.00 |

---

## meta-llama/Llama-3.3-70B-Instruct
### N=8192 K=2048

| batch_size | fp8-tensor-w-tensor-a-noquant (Before) | fp8-tensor-w-tensor-a-noquant (After) | speedup |
|-----------:|---------------------------------------:|--------------------------------------:|----------:|
|          1 |                                   5.65 |                                  6.33 |      1.12 |
|         16 |                                  90.38 |                                106.86 |      1.18 |
|         32 |                                 178.05 |                                195.24 |      1.10 |
|         64 |                                 345.92 |                                467.90 |      1.35 |
|        128 |                                 654.82 |                                861.29 |      1.31 |
|        192 |                                 943.38 |                                935.16 |      0.99 |
|        256 |                                1238.14 |                               1250.33 |      1.01 |

---

## meta-llama/Llama-3.3-70B-Instruct
### N=8192 K=7168

| batch_size | fp8-tensor-w-tensor-a-noquant (Before) | fp8-tensor-w-tensor-a-noquant (After) | speedup |
|-----------:|---------------------------------------:|--------------------------------------:|----------:|
|          1 |                                   9.84 |                                 12.38 |      1.26 |
|         16 |                                 156.65 |                                199.27 |      1.27 |
|         32 |                                 309.39 |                                451.10 |      1.46 |
|         64 |                                 608.19 |                                938.37 |      1.54 |
|        128 |                                1158.40 |                               1777.43 |      1.53 |
|        192 |                                1502.84 |                               1582.51 |      1.05 |
|        256 |                                2021.96 |                               2042.31 |      1.01 |


Raw results:
```
Before:
meta-llama/Llama-3.1-8B-Instruct, N=6144 K=4096, BF16 vs FP8 GEMMs TFLOP/s:                                                           
BF16 vs FP8 GEMMs:                                                                                                                         
…